Mark tests accessing online resources as "Online", make sure tests that do not need to hit external servers do not do so by accident
