+++ modules/extension/xsd/xsd-gml3/src/test/java/org/geotools/gml3/GML3EncodingTest.java
@@ -80,98 +80,6 @@ public class GML3EncodingTest extends TestCase {
         XMLUnit.setXpathNamespaceContext(new SimpleNamespaceContext(namespaces));
     }
 
-    boolean isOffline() throws Exception {
-        // this test will only run if network is available
-        URL url = new URL("http://schemas.opengis.net");
-
-        try {
-            URLConnection conn = url.openConnection();
-            conn.setConnectTimeout(2000);
-            conn.getInputStream().read();
-        } catch (Exception e) {
-            return true;
-        }
-
-        return false;
-    }
-
-    public void testWithConfiguration() throws Exception {
-        if (isOffline()) {
-            return;
-        }
-
-        TestConfiguration configuration = new TestConfiguration();
-
-        // first parse in test data
-        Parser parser = new Parser(configuration);
-        SimpleFeatureCollection fc = (SimpleFeatureCollection) parser.parse(TestConfiguration.class
-                .getResourceAsStream("test.xml"));
-        assertNotNull(fc);
-
-        XSDSchema schema = TEST.getInstance().getSchema();
-        assertNotNull(schema);
-
-        Encoder encoder = new Encoder(configuration, schema);
-        ByteArrayOutputStream output = new ByteArrayOutputStream();
-        encoder.write(fc, TEST.TestFeatureCollection, output);
-
-        validate(output.toByteArray(), configuration);
-    }
-
-    public void testWithApplicationSchemaConfiguration() throws Exception {
-        if (isOffline()) {
-            return;
-        }
-
-        // The schema location needs to be a well formed URI/URL, a file path is not sufficient.
-        String schemaLocation = TestConfiguration.class.getResource("test.xsd").toString();
-
-        ApplicationSchemaConfiguration configuration = new ApplicationSchemaConfiguration(
-                TEST.NAMESPACE, schemaLocation);
-
-        // first parse in test data
-        Parser parser = new Parser(configuration);
-        SimpleFeatureCollection fc = (SimpleFeatureCollection) parser.parse(TestConfiguration.class
-                .getResourceAsStream("test.xml"));
-        assertNotNull(fc);
-
-        XSDSchema schema = TEST.getInstance().getSchema();
-        assertNotNull(schema);
-
-        Encoder encoder = new Encoder(configuration, schema);
-
-        ByteArrayOutputStream output = new ByteArrayOutputStream();
-        encoder.write(fc, TEST.TestFeatureCollection, output);
-
-        validate(output.toByteArray(), configuration);
-    }
-
-    void validate(byte[] data, Configuration configuration) throws Exception {
-        SchemaFactory sf = SchemaFactory.newInstance(XMLConstants.W3C_XML_SCHEMA_NS_URI);
-
-        Schema s = sf.newSchema(new URI(configuration.getSchemaFileURL()).toURL());
-
-        Validator v = s.newValidator();
-
-        final ArrayList errors = new ArrayList();
-        DefaultHandler handler = new DefaultHandler() {
-            public void error(SAXParseException e) throws SAXException {
-                System.out.println(e.getMessage());
-                errors.add(e);
-            }
-
-            public void fatalError(SAXParseException e) throws SAXException {
-                System.out.println(e.getMessage());
-                errors.add(e);
-            }
-        };
-
-        v.setErrorHandler(handler);
-        v.validate(new StreamSource(new ByteArrayInputStream(data)));
-
-        assertTrue(errors.isEmpty());
-    }
-
     public void testEncodeFeatureWithBounds() throws Exception {
         SimpleFeature feature = GML3MockData.feature();
         TestConfiguration configuration = new TestConfiguration();
