+++ modules/unsupported/wfs-ng/src/test/resources/org/geotools/data/wfs/test-data/CubeWerx_nsdi/1.1.0/GetFeature_RoadSeg.xml
@@ -8,7 +8,7 @@
   xmlns:wfs="http://www.opengis.net/wfs"
   xmlns:gml="http://www.opengis.net/gml"
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://www.fgdc.gov/framework/073004/transportation http://frameworkwfs.usgs.gov/framework/wfs/wfs.cgi?DATASTORE=Framework&amp;service=WFS&amp;version=1.1.0&amp;request=DescribeFeatureType&amp;outputFormat=GML3&amp;typeName=ROADSEG  http://www.fgdc.gov/framework/073004 http://frameworkwfs.usgs.gov/framework/schemas/gmlsf1/Framework.xsd http://www.opengis.net/wfs http://schemas.opengis.net/wfs/1.1.0/wfs.xsd http://www.opengis.net/gml http://schemas.opengis.net/gml/3.1.1/base/feature.xsd">
+  xsi:schemaLocation="http://www.fgdc.gov/framework/073004/transportation ./DescribeFetaureType_RoadSeg.xsd">
   <gml:featureMember>
     <trans:RoadSeg gml:id="CWFID.ROADSEG.0.0">
       <trans:lastUpdateDate>2005-06-01T00:00:00Z</trans:lastUpdateDate>
